+++ include/rtabmap_ros/CoreWrapper.h
