+++ include/rtabmap_ros/MapsManager.h
