+++ include/rtabmap_ros/PreferencesDialogROS.h
